add examples plots and update readme
